+++ src/main/scala/org/podval/docbook/gradle/DocBookExtension.scala
@@ -21,6 +21,9 @@ class DocBookExtension(project: Project) {
   @BeanProperty val outputFormats: ListProperty[String] =
     project.getObjects.listProperty(classOf[String])
 
+  @BeanProperty val outputFormats2: ListProperty[String] =
+    project.getObjects.listProperty(classOf[String])
+
   @BeanProperty val epubEmbeddedFonts: ListProperty[String] =
     project.getObjects.listProperty(classOf[String])
 }