+++ src/main/scala/org/podval/docbook/gradle/Util.scala
@@ -15,6 +15,9 @@ object Util {
   def drop(what: String, from: String): Option[String] =
     if (from.startsWith(what)) Some(from.drop(what.length)) else None
 
+  def subdirectory(directory: File, subdirectoryName: Option[String]): File =
+    subdirectoryName.fold(directory)(new File(directory, _))
+
   // Maybe in JDK 9 and later I won't need to deal with '$'?
   // see https://bugs.java.com/bugdatabase/view_bug.do?bug_id=JDK-8057919
   def className(obj: AnyRef): String = obj.getClass.getSimpleName.replace("$", "")