+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- XSLT 2.0 DocBook stylesheets with Saxon 9 HE
 
 ## [0.2.1] - 02.06.2019
 - EPUB3